+++ app/(map)/_components/Map.tsx
@@ -12,9 +12,9 @@ import {
   CarouselPrevious,
 } from "@/app/_common/shadcn/ui/carousel";
 
+import WithSearchTokens from "@/app/_common/components/WithSearchTokens";
 import MapComponent from "@/app/_common/components/MapComponent";
 
-import WithSearchTokens from "@/app/_common/components/WithSearchTokens";
 
 import REGION_CODE from "@/app/_common/constants/regionCode";
